+++ Modules/SWMB/LocalMachine-Application.psm1
@@ -1016,5 +1016,28 @@ Function TweakViewStorageSenseTrashCleanup { # RESINFO
 
 ################################################################
 
+Function TweakDisableWindowsCopilot { # RESINFO
+	Write-Output "Turn Off Windows Copilot..."
+	If (!(Test-Path "HKLM\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot")) {
+		New-Item -Path  "HKLM\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot" -Force | Out-Null
+	}
+	Set-ItemProperty -Path  "HKLM\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot" -Name "TurnOffWindowsCopilot" -Type DWord -Value 1
+}
+
+# Enable Copilot
+Function TweakEnableWindowsCopilot { # RESINFO
+	Write-Output "Dafault setup for Windows Copilot..."
+	Remove-ItemProperty -Path "HKLM\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot" -Name "TurnOffWindowsCopilot" -ErrorAction SilentlyContinue
+}
+
+# View Copilot
+Function TweakViewWindowsCopilot { # RESINFO
+	Write-Output "View Windows Copilot"
+	Get-ItemProperty -Path "HKLM\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot" -Name "TurnOffWindowsCopilot" -ErrorAction SilentlyContinue
+}
+
+################################################################
+
+
 # Export functions
 Export-ModuleMember -Function *