+++ NEWS.md
@@ -21,6 +21,7 @@ SWLN defaults to the latest available version of SWMB.
 Add link to the [Harden Community](https://hardenad.net/) in [REFERENCES](REFERENCES.md).
 
 New presets/rules:
+* 2024/02/19 - `DisableWindowsCopilot` -> Disable Windows Copilot
 * 2024/02/19 - `DisableOneSettingsDownloads` -> Windows will not download configuration settings from the OneSettings service
 * 2024/02/19 - `DisableDiagnosticLogs` -> Diagnostic logs collected
 * 2024/02/07 - `UninstallHPWolfSecurity` -> Uninstall HP Wolf Security